// ui/src/lib/api/generated/model/moodboardSummary.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Centoire API
 * AI-powered fashion platform API
 * OpenAPI spec version: 1.0.0
 */
import type { MoodboardSummaryStatus } from './moodboardSummaryStatus';
import type { MoodboardSummaryInputMode } from './moodboardSummaryInputMode';

export interface MoodboardSummary {
  _id: string;
  userId: string;
  jobId: string;
  status: MoodboardSummaryStatus;
  inputMode: MoodboardSummaryInputMode;
  name: string;
  createdAt: string;
  updatedAt: string;
}